Drop leanback implementation of search
